Rename pkgs in pm exp, add ping to ipc client
